+++ app/src/main/java/com/chat/joycom/model/Member.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Member(
-    @field:Json(name = "UserId") val userId: Int,
+    @field:Json(name = "UserId") val userId: Long,
     @field:Json(name = "AccountName") val accountName: String,
     @field:Json(name = "Nickname") val nickname: String,
     @field:Json(name = "Avatar") val avatar: String,
